+++ tests/Move/StartOfMonthTest.php
@@ -1,25 +1,25 @@
 <?php
 declare(strict_types = 1);
 
-namespace Tests\Innmind\TimeContinuum\Earth\Move;
+namespace Tests\Innmind\TimeContinuum\Move;
 
 use Innmind\TimeContinuum\{
-    Earth\Move\StartOfDay,
+    Move\StartOfMonth,
     PointInTime,
     Format,
 };
 use PHPUnit\Framework\TestCase;
 
-class StartOfDayTest extends TestCase
+class StartOfMonthTest extends TestCase
 {
     /**
      * @dataProvider cases
      */
     public function testInterface($time, $expected)
     {
-        $startOfDay = new StartOfDay;
+        $startOfMonth = new StartOfMonth;
 
-        $point = $startOfDay(PointInTime::at($time));
+        $point = $startOfMonth(PointInTime::at($time));
 
         $this->assertSame($expected, $point->format(Format::of('Y-m-d H:i:s.u')));
     }
@@ -27,10 +27,9 @@ class StartOfDayTest extends TestCase
     public static function cases(): array
     {
         return [
-            ['2016-02-29 13:12:11.675', '2016-02-29 00:00:00.000000'],
-            ['2018-04-28 01:12:11.675', '2018-04-28 00:00:00.000000'],
+            ['2016-02-29 13:12:11.675', '2016-02-01 00:00:00.000000'],
+            ['2018-04-28 01:12:11.675', '2018-04-01 00:00:00.000000'],
             ['2018-04-01 00:00:00.000', '2018-04-01 00:00:00.000000'],
-            ['2018-11-30T12:13:14.000+0100', '2018-11-30 00:00:00.000000'],
         ];
     }
 }